Increase number of attempts and timeout when delete a dataset

+++ pipelines/ingest-gbif/src/main/java/org/gbif/pipelines/ingest/options/EsPipelineOptions.java
@@ -62,13 +62,13 @@ public interface EsPipelineOptions extends PipelineOptions {
   void setIndexNumberReplicas(Integer indexNumberReplicas);
 
   @Description("ES empty delete index query timeout in seconds")
-  @Default.Integer(3)
+  @Default.Integer(5)
   Integer getSearchQueryTimeoutSec();
 
   void setSearchQueryTimeoutSec(Integer searchQueryTimeoutSec);
 
   @Description("ES empty index query attempts")
-  @Default.Integer(100)
+  @Default.Integer(200)
   Integer getSearchQueryAttempts();
 
   void setSearchQueryAttempts(Integer searchQueryAttempts);
